New authorization idea for login with configuration: * Assuption 1: Database has users of type (<configuration prefix>,<configuration hash>,<description>). 	-These users have usernames similar to database/configuration prefixes, 	-These users have passwords equal to MD5 hash of prefix configuration file, 	-These users have all the necessary permissions specified in remarks database, 	-These users have correct permissions in database for aggregated prefixes.
* Remarks server code was not modified a lot. Right now it just supports login with
"login@prefix" type of username. The standard login procedure will not login this type of
user unless it exists in database. We want to keep "login" of user to use it for draw_sets,
remarks and user params he adds, but we want to login all users of type "x@prefix", "y@prefix",
"z@prefix" to the same database accout called "prefix". Thats why remarks server now parses
the login after incorrect authorization in search of this type of username (it lets the database
have just one user "prefix" for all users of this prefix.

* Draw3: Username is now taken from wx enviroment (wxString wxGetUserName).

@TODO: Implementing hash history (right now every local, uncommited change to
configuration will result in disabling cfglogin).