QUICK FIX: Reading UTF-8 configuration for SCC Menu
That's a quick fix for bad displaying of SCC Menu, when 'szarp.cfg' is
UTF-8 encoded. The problem is deeper (file 'libSzarp/conversion.cc') and
"fixing" it without thinking through may broke many parts of Szarp. This
fix should be safe for older configurations (Win/CP1250 and ISO8859-2)
and only applies the situation when system encoding is "UTF-8".

Committing to 'master', cause it's urgent. Have an eye on it!