New conversion routines that omits invalid characters
+ functions ascii2utf(), local2szarp() and local2utf() have new variant
(with '_ign' suffix), which omits invalid characters (in the input
encoding) and ignore characters that cannot be represented (in the
target encoding).

+ functions A2U(), L2S(), L2U() get additional argument named
'fallback'. If it's set to "true", in case of conversion failure (due
to invalid multibyte sequence) function catches an exception and tries
to repeat conversion using '_ign' function variant. 'fallback' defaults
to "false", so existing usages will preserve behavior.

+ set 'fallback' to "true" in L2S(), L2U(), A2U() calls, where input
encoding is uncertain.

+ changed A2U() calls to L2U(), where proper.