*sz4: a slight format change
Change of the way time is encoded. Instead of using absolute values per
each data point, time is encoded as a variable width increment since the time
of a previous probe.

This is only meaner part of the change. Change to the libSzarp2 will follow.

also: missing file heartbeat.py