make libtool to link libSzarp* libs statically
It is kind of micro-optimisation for 32bit but it was simply annoying
that according to profiler most of the time in sz4 was spent on
"thunk" - piece of code that is copying PC register value to something
that 32bit processor can use for relative addressing. This "thunk"
is something that is only used on 32bit builds.