boruta modbus rtu timers fix
Recent changes that introduced more elobarate timers handling
also introduced a bug. The bug is that if 1.5 Tc timer fires
immidiately after first character of a response is received,
modbus code will calculate crc over the previous packet it got
(i.e.  response to a previous request) and if that matches, that
previous response will be consumed as if it was response to
current request.