borutadmn: modbus tcp - prev change followup/bugfix
If the response contains invalid transaction id just ignore it, instead
of progressing with queries. Ignoring such request makes more sense
because usually that happens when a device is lagging behind with
its responses, so it makes sense to wait.

This also fixes a bug where response with wrong transaction id was leaving
deadliner timer active, which could lead to a crash.

Also - don't do anything on deadline timer tick if the unit is
in idle state.

One day we will have a unit test for this code...