fix: Improper memomory management for wxLog pointers used by wxLog::SetActiveTarget.
wxLog::SetActiveTarget is a static method keeping a pointer to wxLog passed as argument.
It returns the previous wxLog pointer while being called. To avoid memory leaks the
previous pointer returned should be deleted and on App exit NULL should be passed
as argument. SCCMenu should not be an owner of wxLogStderr since wxLog::SetActiveTarget
is used throughout the application (there will be a segfault).

(FYI: https://wiki.wxwidgets.org/WxLog)