feature: S7Daemon: Better handling of min/max configuration.
This method uses SZARP style of converting floating point to integer
using configured precision. This guarantees that min/max configuration
allows values to be written that are what SZARP users expect to
appear in database. Epsilon and more robust methods can be used for
Sz4 as a separate utility in libSzarp2. Unit tests are also
included.